+++ src/main/java/org/tangxi/testcase/execution/mapper/TestCaseMapper.java
@@ -3,5 +3,5 @@ package org.tangxi.testcase.execution.mapper;
 import org.tangxi.testcase.execution.model.TestCase;
 
 public interface TestCaseMapper {
-    TestCase selectTestCaseById(Long id);
+    TestCase selectTestCaseById(int id);
 }
